incorporate feedback from review
