+++ rpc/master/hosts_client.go
@@ -15,8 +15,6 @@ package master
 
 import (
 	"github.com/control-center/serviced/domain/host"
-	"github.com/control-center/serviced/volume"
-	"github.com/zenoss/glog"
 )
 
 //GetHost gets the host for the given hostID or nil
@@ -66,17 +64,6 @@ func (c *Client) FindHostsInPool(poolID string) ([]host.Host, error) {
 	response := make([]host.Host, 0)
 	if err := c.call("FindHostsInPool", poolID, &response); err != nil {
 		return []host.Host{}, err
-	}
-	return response, nil
-}
-
-//GetVolumeStatus gets status information for the given volume or nil
-func (c *Client) GetVolumeStatus(volumeNames []string) (*volume.Statuses, error) {
-	glog.V(2).Infof("[hosts_client.go]master.GetVolumeStatus(%v)", volumeNames)
-	response := &volume.Statuses{}
-	if err := c.call("GetVolumeStatus", volumeNames, response); err != nil {
-		glog.V(2).Infof("\tcall to GetVolumeStatus returned error: %v", err)
-		return nil, err
 	}
 	return response, nil
 }
