+++ rpc/master/hosts_server.go
@@ -17,8 +17,6 @@ import (
 	"github.com/control-center/serviced/domain/host"
 
 	"errors"
-	"github.com/control-center/serviced/volume"
-	"github.com/zenoss/glog"
 )
 
 // GetHost gets the host
@@ -76,17 +74,5 @@ func (s *Server) FindHostsInPool(poolID string, hostReply *[]host.Host) error {
 		return err
 	}
 	*hostReply = hosts
-	return nil
-}
-
-// GetVolumeStatus gets the volume status
-func (s *Server) GetVolumeStatus(volumeNames []string, reply *volume.Statuses) error {
-	glog.V(2).Infof("[hosts_server.go]master.GetVolumeStatus(%v, %v)\n", volumeNames, reply)
-	response := volume.GetStatus(volumeNames)
-	if response == nil {
-		glog.V(2).Infof("\tCall to volume.getStatus failed: (%v). Returning error.", response)
-		return errors.New("hosts_server.go host not found")
-	}
-	*reply = *response
 	return nil
 }
